dmkod-integration-app/requirements.txt
flask
psycopg2-binary
python-dotenv
pandas
openpyxl
requests
gunicorn
Flask-Login
Flask-WTF
bcrypt